chore: move renderToString to vue-compat
