+++ src/util/__tests__/createServerRootMixin.test.js
@@ -2,17 +2,20 @@ import { mount, createSSRApp } from '../../../test/utils';
 import Router from 'vue-router';
 import Vuex from 'vuex';
 import { createStore } from 'vuex4';
-import {
-  createServerRootMixin,
-  renderToString,
-} from '../createServerRootMixin';
+import { createServerRootMixin } from '../createServerRootMixin';
 import InstantSearchSsr from '../../components/InstantSearchSsr';
 import Configure from '../../components/Configure';
 import SearchBox from '../../components/SearchBox.vue';
 import { createWidgetMixin } from '../../mixins/widget';
 import { createFakeClient } from '../testutils/client';
 import { createSerializedState } from '../testutils/helper';
-import { isVue3, isVue2, Vue2, renderCompat } from '../vue-compat';
+import {
+  isVue3,
+  isVue2,
+  Vue2,
+  renderCompat,
+  renderToString,
+} from '../vue-compat';
 import {
   SearchResults,
   SearchParameters,
